# perf(schema): use FnvHashMap for type and directive names