--- a/juniper/Cargo.toml
+++ b/juniper/Cargo.toml
@@ -25,6 +25,7 @@ expose-test-schema = []
 default = ["url", "uuid"]
 
 [dependencies]
+fnv = "1.0.3"
 ordermap = { version = "^0.2.11", features = ["serde-1"] }
 serde = { version = "^1.0.8" }
 serde_derive = {version="^1.0.8" }