--- a/juniper/src/lib.rs
+++ b/juniper/src/lib.rs
@@ -122,6 +122,7 @@ extern crate serde_derive;
 #[cfg(any(test, feature = "expose-test-schema"))]
 extern crate serde_json;
 
+extern crate fnv;
 extern crate ordermap;
 
 #[cfg(any(test, feature = "url"))]