Changes `SchemaType` to use a `HashMap` backed by the [https://github.com/servo/rust-fnv](Fowler–Noll–Vo hash function). This hash function performs much better for smaller inputs (think short strings like most GraphQL type names) than the default hash function, SipHash. The downside is that it does not have the same cryptographic properties, which open it up to DoS vulnerabilities when the input can be manipulated from the outside.

In this change set, I have chosen to apply FNV only to the type and directive names tracked by the `SchemaType`, for two reasons:

1. Look-ups in `SchemaType::type_by_name` and co. have been showing up as the most frequent/expensive use of hashing for my workload.
2. I believe it is reasonable to assume that a malicious party would not have control over the schema type, therefore could not leverage FNV for a DoS attack (contrary to e.g. using it in the validation logic).

Here you can see how `SchemaType::type_by_name`, `::concrete_type_by_name` as well as both `::make_type` and `ValidatorContext::with_pushed_type` (these use `type_by_name` internally) show up in my profile:

![schematype-hash-perf](https://user-images.githubusercontent.com/2677165/31069231-424b4608-a7a7-11e7-8cd1-8ca9738d7be2.jpg)

In my (somewhat limited) benchmark, the change appears to lead to a worthwhile performance improvement.

Pre:

```
Running 30s test @ http://localhost:3000/graphql
  2 threads and 4 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   572.58us  336.26us  14.93ms   97.35%
    Req/Sec     3.54k   284.33     3.93k    77.08%
  Latency Distribution
     50%  526.00us
     75%  573.00us
     90%  663.00us
     99%    1.15ms
  211827 requests in 30.10s, 26.46MB read
Requests/sec:   7037.39
Transfer/sec:      0.88MB
```

Post:

```
Running 30s test @ http://localhost:3000/graphql
  2 threads and 4 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   550.45us  213.25us   8.73ms   97.53%
    Req/Sec     3.64k   169.32     3.93k    79.87%
  Latency Distribution
     50%  520.00us
     75%  554.00us
     90%  616.00us
     99%    0.96ms
  217928 requests in 30.10s, 27.23MB read
Requests/sec:   7240.40
Transfer/sec:      0.90MB
```