@Techcable 2017-10-07
Why not seahash? Isn't that [signifigantly faster](https://docs.rs/seahash/3.0.5/seahash/#benchmark)?