@srijs 2017-10-08
# Juniper Benchmarks

## FnvHasher

```
running 2 tests
test introspection_query ... bench:   1,744,556 ns/iter (+/- 422,545)
test query_type_name     ... bench:      14,961 ns/iter (+/- 3,658)

test result: ok. 0 passed; 0 failed; 0 ignored; 2 measured
```

## SeaHasher

```
running 2 tests
test introspection_query ... bench:   1,794,380 ns/iter (+/- 554,186)
test query_type_name     ... bench:      15,851 ns/iter (+/- 3,547)

test result: ok. 0 passed; 0 failed; 0 ignored; 2 measured
```

## DefaultHasher

```
running 2 tests
test introspection_query ... bench:   2,058,095 ns/iter (+/- 358,108)
test query_type_name     ... bench:      18,215 ns/iter (+/- 6,773)

test result: ok. 0 passed; 0 failed; 0 ignored; 2 measured
```