@mhallin 2017-10-08
I think this looks good, thanks for the improvement! I think your assumptions about hash-collisions attacks are reasonable.

Without having done any deeper performance analysis, I'd think that Juniper's biggest overhead on larger queries is running the validation passes. Having a MFU/MRU cache of parsed and validated ASTs could probably help there, or implementing [Apollo's persisted queries](https://dev-blog.apollodata.com/persisted-graphql-queries-with-apollo-client-119fd7e6bba5).