@srijs 2017-10-08
@mhallin agreed. Coincidentally, my work on removing the lifetime parameter from the AST for purposes of introducing async could prove useful for introducing a query cache!